fix: apply review fixes for 3.0 release
