+++ src/workflow-script.ts
@@ -1,0 +1,12 @@
+export type ParsedWorkflowMeta = {
+  name: string;
+  description: string;
+  [k: string]: unknown;
+};
+
+export {
+  parseWorkflow,
+  makeGuardedDate,
+  makeGuardedMath,
+  workflowStringify,
+} from "./workflow-script.mjs";
